+++ Flamy2D/Audio/Sound.cs
@@ -1,16 +1,12 @@
 ﻿using Flamy2D.Assets;
 using NVorbis;
-using OpenTK.Audio;
 using OpenTK.Audio.OpenAL;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading;
 
 namespace Flamy2D.Audio
 {
-    public class Sound : IDisposable
+    public class Sound : IDisposable, IAsset
     {
         const int DefaultBufferCount = 3;
 
@@ -112,7 +108,7 @@ namespace Flamy2D.Audio
 
             Preparing = false;
 
-            OggStreamer.Instance.AddStream(this);
+            AudioDevice.Instance.AddStream(this);
         }
 
         public void Pause()
@@ -120,7 +116,7 @@ namespace Flamy2D.Audio
             if (AL.GetSourceState(alSourceId) != ALSourceState.Playing)
                 return;
 
-            OggStreamer.Instance.RemoveStream(this);
+            AudioDevice.Instance.RemoveStream(this);
             AL.SourcePause(alSourceId);
             ALHelper.Check();
         }
@@ -130,7 +126,7 @@ namespace Flamy2D.Audio
             if (AL.GetSourceState(alSourceId) != ALSourceState.Paused)
                 return;
 
-            OggStreamer.Instance.AddStream(this);
+            AudioDevice.Instance.AddStream(this);
             AL.SourcePlay(alSourceId);
             ALHelper.Check();
         }
@@ -146,7 +142,7 @@ namespace Flamy2D.Audio
             lock (stopMutex)
             {
                 NotifyFinished();
-                OggStreamer.Instance.RemoveStream(this);
+                AudioDevice.Instance.RemoveStream(this);
             }
         }
 
@@ -186,7 +182,7 @@ namespace Flamy2D.Audio
 
             lock (prepareMutex)
             {
-                OggStreamer.Instance.RemoveStream(this);
+                AudioDevice.Instance.RemoveStream(this);
 
                 if (state != ALSourceState.Initial)
                     Empty();
@@ -262,12 +258,12 @@ namespace Flamy2D.Audio
             if (precache)
             {
                 // Fill first buffer synchronously
-                OggStreamer.Instance.FillBuffer(this, alBufferIds[0]);
+                AudioDevice.Instance.FillBuffer(this, alBufferIds[0]);
                 AL.SourceQueueBuffer(alSourceId, alBufferIds[0]);
                 ALHelper.Check();
 
                 // Schedule the others asynchronously
-                OggStreamer.Instance.AddStream(this);
+                AudioDevice.Instance.AddStream(this);
             }
 
             Ready = true;
@@ -284,226 +280,5 @@ namespace Flamy2D.Audio
         }
     }
 
-    public class OggStreamer : IDisposable
-    {
-        const float DefaultUpdateRate = 10;
-        const int DefaultBufferSize = 44100;
-
-        static readonly object singletonMutex = new object();
-
-        readonly object iterationMutex = new object();
-        readonly object readMutex = new object();
-
-        readonly float[] readSampleBuffer;
-        readonly short[] castBuffer;
-
-        readonly HashSet<OggStream> streams = new HashSet<OggStream>();
-        readonly List<OggStream> threadLocalStreams = new List<OggStream>();
-
-        Thread underlyingThread;
-        volatile bool cancelled;
-
-        public float UpdateRate { get; private set; }
-        public int BufferSize { get; private set; }
-
-        public ILogger Logger { private get; set; }
-
-        static OggStreamer instance;
-        public static OggStreamer Instance
-        {
-            get
-            {
-                lock (singletonMutex)
-                {
-                    if (instance == null)
-                        throw new InvalidOperationException("No instance running");
-                    return instance;
-                }
-            }
-            private set { lock (singletonMutex) instance = value; }
-        }
-
-        /// <summary>
-        /// Constructs an OggStreamer that plays ogg files in the background
-        /// </summary>
-        /// <param name="bufferSize">Buffer size</param>
-        /// <param name="updateRate">Number of times per second to update</param>
-        /// <param name="internalThread">True to use an internal thread, false to use your own thread, in which case use must call EnsureBuffersFilled periodically</param>
-        public OggStreamer(int bufferSize = DefaultBufferSize, float updateRate = DefaultUpdateRate, bool internalThread = true)
-        {
-            lock (singletonMutex)
-            {
-                if (instance != null)
-                    throw new InvalidOperationException("Already running");
-
-                Instance = this;
-                if (internalThread)
-                {
-                    underlyingThread = new Thread(EnsureBuffersFilled) { Priority = ThreadPriority.Lowest };
-                    underlyingThread.Start();
-                }
-                else
-                {
-                    // no need for this, user is in charge
-                    updateRate = 0;
-                }
-            }
-
-            UpdateRate = updateRate;
-            BufferSize = bufferSize;
-
-            readSampleBuffer = new float[bufferSize];
-            castBuffer = new short[bufferSize];
-
-            Logger = NullLogger.Default;
-        }
-
-        public void Dispose()
-        {
-            lock (singletonMutex)
-            {
-                Debug.Assert(Instance == this, "Two instances running, somehow...?");
-
-                cancelled = true;
-                lock (iterationMutex)
-                    streams.Clear();
-
-                Instance = null;
-                underlyingThread = null;
-            }
-        }
-
-        internal bool AddStream(OggStream stream)
-        {
-            lock (iterationMutex)
-                return streams.Add(stream);
-        }
-        internal bool RemoveStream(OggStream stream)
-        {
-            lock (iterationMutex)
-                return streams.Remove(stream);
-        }
-
-        public bool FillBuffer(OggStream stream, int bufferId)
-        {
-            int readSamples;
-            lock (readMutex)
-            {
-                readSamples = stream.Reader.ReadSamples(readSampleBuffer, 0, BufferSize);
-                CastBuffer(readSampleBuffer, castBuffer, readSamples);
-            }
-            AL.BufferData(bufferId, stream.Reader.Channels == 1 ? ALFormat.Mono16 : ALFormat.Stereo16, castBuffer,
-                          readSamples * sizeof(short), stream.Reader.SampleRate);
-            ALHelper.Check();
-
-            if (readSamples == BufferSize) Logger.Log(LogEvent.NewPacket, stream);
-            else Logger.Log(LogEvent.LastPacket, stream);
-            Logger.Log(LogEventSingle.MemoryUsage, () => GC.GetTotalMemory(true));
-
-            return readSamples != BufferSize;
-        }
-        public static void CastBuffer(float[] inBuffer, short[] outBuffer, int length)
-        {
-            for (int i = 0; i < length; i++)
-            {
-                var temp = (int)(32767f * inBuffer[i]);
-                if (temp > short.MaxValue) temp = short.MaxValue;
-                else if (temp < short.MinValue) temp = short.MinValue;
-                outBuffer[i] = (short)temp;
-            }
-        }
-
-        public void EnsureBuffersFilled()
-        {
-            do
-            {
-                threadLocalStreams.Clear();
-                lock (iterationMutex) threadLocalStreams.AddRange(streams);
-
-                foreach (var stream in threadLocalStreams)
-                {
-                    lock (stream.prepareMutex)
-                    {
-                        lock (iterationMutex)
-                            if (!streams.Contains(stream))
-                                continue;
-
-                        bool finished = false;
-
-                        int queued;
-                        AL.GetSource(stream.alSourceId, ALGetSourcei.BuffersQueued, out queued);
-                        ALHelper.Check();
-                        int processed;
-                        AL.GetSource(stream.alSourceId, ALGetSourcei.BuffersProcessed, out processed);
-                        ALHelper.Check();
-
-                        if (processed == 0 && queued == stream.BufferCount) continue;
-
-                        int[] tempBuffers;
-                        if (processed > 0)
-                            tempBuffers = AL.SourceUnqueueBuffers(stream.alSourceId, processed);
-                        else
-                            tempBuffers = stream.alBufferIds.Skip(queued).ToArray();
-
-                        int bufIdx = 0;
-                        for (; bufIdx < tempBuffers.Length; bufIdx++)
-                        {
-                            finished |= FillBuffer(stream, tempBuffers[bufIdx]);
-
-                            if (finished)
-                            {
-                                if (stream.IsLooped)
-                                {
-                                    stream.Reader.DecodedTime = TimeSpan.Zero;
-                                    if (bufIdx == 0)
-                                    {
-                                        // we didn't have any buffers left over, so let's start from the beginning on the next cycle...
-                                        continue;
-                                    }
-                                }
-                                else
-                                {
-                                    lock (stream.stopMutex)
-                                    {
-                                        stream.NotifyFinished();
-                                    }
-                                    streams.Remove(stream);
-                                    break;
-                                }
-                            }
-                        }
-
-                        AL.SourceQueueBuffers(stream.alSourceId, bufIdx, tempBuffers);
-                        ALHelper.Check();
-
-                        if (finished && !stream.IsLooped)
-                            continue;
-                    }
-
-                    lock (stream.stopMutex)
-                    {
-                        if (stream.Preparing) continue;
-
-                        lock (iterationMutex)
-                            if (!streams.Contains(stream))
-                                continue;
-
-                        var state = AL.GetSourceState(stream.alSourceId);
-                        if (state == ALSourceState.Stopped)
-                        {
-                            Logger.Log(LogEvent.BufferUnderrun, stream);
-                            AL.SourcePlay(stream.alSourceId);
-                            ALHelper.Check();
-                        }
-                    }
-                }
-
-                if (UpdateRate > 0)
-                {
-                    Thread.Sleep((int)(1000 / UpdateRate));
-                }
-            }
-            while (underlyingThread != null && !cancelled);
-        }
-    }
+    
 }
