+++ Flamy2D/Flamy2D.csproj
@@ -111,6 +111,7 @@
     <Compile Include="Assets\Providers\TextureProvider.cs" />
     <Compile Include="Audio\ALException.cs" />
     <Compile Include="Audio\ALHelper.cs" />
+    <Compile Include="Audio\AudioDevice.cs" />
     <Compile Include="Audio\Sound.cs" />
     <Compile Include="Buffer\GLBuffer.cs" />
     <Compile Include="Buffer\GLBufferDynamic.cs" />
